requirements.txt
h5py
open3d>=0.12.0,<=0.14.1
pykitti
opencv-python
pytorch_metric_learning
cython
wandb